requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
deepface>=0.0.79
tf-keras>=2.15.0
speechbrain>=0.5.16
librosa>=0.10.1
soundfile>=0.12.1
numpy>=1.24.0
pillow>=10.0.0
opencv-python-headless>=4.8.0
pydub>=0.25.1
